backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
pydantic
python-jose
passlib
email-validator
bcrypt==4.2.0